reorganized files for testings and updated README
