+++ vitest.config.ts
@@ -5,7 +5,7 @@ export default defineConfig({
   test: {
     environment: "node",
     globals: false,
-    setupFiles: ["./tests/setup.ts"],
+    setupFiles: ["./tests/lib/support/setup.ts"],
     fileParallelism: false,
   },
   resolve: {
